// src/app/models/staff-expense.ts
import {Staff} from "./staff";

export class StaffExpense {
  staff_expense_uuid: string;
  supplier: string;
  category: string;
  categoryLbl: string;
  purchase_date: string;
  total_amount: string;
  currency: string;
  vat: number;
  reimbursable: number;
  description: string;
  file: string;
  status: number;
  statusLbl: string;
  staff_id: string;
  created_at: string;
  updated_at: string;
  created_by: string;
  updated_by: string;
  staff: Staff;
}